create readme.md
Bubblesort-numbers is a number sorting algorithm where the user is prompted to enter 10 positive or negative integers. After the user had entered the integer values their input will be shown in tabular and ascending order, however unsorted. To sort the array the numbers are passed by reference to a function called Bubblesort where the numbers are arranged from smallest to greatest value.